feat: add TimetablesType and update OrbiterDag, for SerDe
